+++ Documentation/cmd-plugins.txt
@@ -46,8 +46,8 @@ linknvme:nvme-micron-smart-add-log[1]::
 linknvme:nvme-micron-temperature-stats[1]::
 	Retrieves temperature information of given micron device
 
-linknvme:nvme-micron-ocp-telemetry-log-parse[1]::
-	Parse OCP Telemetry DA1 and DA2 logs.
+linknvme:nvme-ocp-internal-log[1]::
+	Retrieves and parses OCP Telemetry DA1 and DA2 logs.
 
 linknvme:nvme-netapp-ontapdevices[1]::
 	Display information about ONTAP devices
