+++ Documentation/meson.build
@@ -81,7 +81,6 @@ adoc_sources = [
   'nvme-micron-selective-download',
   'nvme-micron-smart-add-log',
   'nvme-micron-temperature-stats',
-  'nvme-micron-ocp-telemetry-log-parse',
   'nvme-netapp-ontapdevices',
   'nvme-netapp-smdevices',
   'nvme-ns-descs',
@@ -105,6 +104,7 @@ adoc_sources = [
   'nvme-ocp-smart-add-log',
   'nvme-ocp-telemetry-string-log-page',
   'nvme-ocp-unsupported-reqs-log-pages',
+  'nvme-ocp-internal-log',
   'nvme-persistent-event-log',
   'nvme-pred-lat-event-agg-log',
   'nvme-predictable-lat-log',
