+++ completions/_nvme
@@ -180,12 +180,18 @@ _nvme () {
 				local _internal_log
 				_internal_log=(
 				/dev/nvme':supply a device to use (required)'
-				--telemetry_type=':Telemetry Type; host (Create bit) or controller'
-				-t':alias for --telemetry_type'
-				--telemetry_data_area=':Telemetry Data Area; 1 or 3'
-				-a':alias for --telemetry_data_area'
+				--telemetry-type=':Telemetry Type; host or controller generated'
+				-t':alias for --telemetry-type'
+				--data-area=':Telemetry Data Area; 1 or 2'
+				-a':alias for --data-area'
 				--output-file=':Output file name with path'
 				-o':alias for --output-file'
+				--telemetry-log=':Telemetry log binary'
+				-l':alias for --telemetry-log'
+				--string-log=':String log binary'
+				-s':alias for --string-log'
+				--output-format':Output format: normal|json'
+				-f':alias for --output-format'
 				)
 				_arguments '*:: :->subcmds'
 				_describe -t commands "nvme ocp internal-log options" _internal_log
@@ -529,29 +535,6 @@ _nvme () {
 				;;
 			esac
 			;;
-		(*)
-			_files
-			;;
-		esac
-			;;
-		(micron)
-			case ${words[2]} in
-			(ocp-telemetry-log-parse)
-				local _ocp-telemetry-log-parse
-				_ocp-telemetry-log-parse=(
-				/dev/nvme':supply a device to use (required)'
-				--output-file=':Output file name with path'
-				-o':alias for --output-file'
-				--telemetry-log=':Telemetry log binary'
-				-l':alias for --telemetry-log'
-				--string-log=':String log binary'
-				-s':alias for --string-log'
-				--format':Output format: normal|json'
-				-f':alias for --format'
-				)
-				_arguments '*:: :->subcmds'
-				_describe -t commands "nvme micron ocp-telemetry-log-parse" _ocp-telemetry-log-parse
-				;;
 		(*)
 			_files
 			;;
