+++ plugins/meson.build
@@ -11,8 +11,6 @@ if json_c_dep.found()
     'plugins/inspur/inspur-nvme.c',
     'plugins/intel/intel-nvme.c',
     'plugins/memblaze/memblaze-nvme.c',
-    'plugins/micron/micron-utils.c',
-    'plugins/micron/micron-ocp-telemetry.c',
     'plugins/micron/micron-nvme.c',
     'plugins/nbft/nbft-plugin.c',
     'plugins/netapp/netapp-nvme.c',
